add some initial serialization tests
